+++ src/ui/components/TranslationGame/EndGame/EndGameHeader.jsx
@@ -1,0 +1,10 @@
+export default function EndGameHeader() {
+  return (
+    <div className="text-center mb-8">
+      <h1 className="text-4xl font-extrabold text-red-600 mb-2">
+        🏁 המשחק נגמר!
+      </h1>
+      <p className="text-lg text-gray-700">תודה על ההשתתפות 🎉</p>
+    </div>
+  );
+}
